beidl message

+++ src/main/java/me/jweissen/aeticket/dto/request/UserUpdateRequestDto.java
@@ -1,4 +1,4 @@
 package me.jweissen.aeticket.dto.request;
 
-public record UserUpdateRequestDto(Integer id, String firstname, String lastname, String email) {
+public record UserUpdateRequestDto(Long id, String firstname, String lastname, String email) {
 }
